requirements.txt
# pip install markdown weasyprint python-docx tkinterdnd2 beautifulsoup4
markdown
weasyprint
python-docx
tkinterdnd2
beautifulsoup4
